requirements.txt
openai==0.27.1
redis==4.5.1
requests==2.28.2
termcolor==2.2.0
jupyter
ipykernel
textract
langchain==0.0.312
python-dotenv==1.0.0
httpx
fastapi~=0.98.0
uvicorn~=0.22.0
pinecone-client==2.2.2
transformers
fastapi-sessions
requests
pydantic~=1.10.12
bs4
sse-starlette